extract common code for a simple chat app, for both chrome and web
