+++ Gruntfile.coffee
@@ -27,7 +27,7 @@ taskManager.add 'samples', [
   'simpleSocksFirefoxApp'
   'copypasteSocks'
   'simpleTurn'
-  'simpleChurnChatChromeApp'
+  'simpleChat'
   'copypasteChurnChatChromeApp'
   'adventure'
   'uprobe'
@@ -42,12 +42,26 @@ taskManager.add 'dist', [
   'copy:dist'
 ]
 
-# Build the simple freedom chat sample app.
-taskManager.add 'simpleFreedomChat', [
+# Simple chat app.
+taskManager.add 'simpleChatBase', [
   'base'
-  'copy:libsForSimpleFreedomChat'
-  'browserify:simpleFreedomChatMain'
-  'browserify:simpleFreedomChatFreedomModule'
+  'browserify:simpleChatFreedomModule'
+  'browserify:simpleChatMain'
+]
+
+taskManager.add 'simpleChatChromeApp', [
+  'simpleChatBase'
+  'copy:libsForSimpleChatChromeApp'
+]
+
+taskManager.add 'simpleChatWebApp', [
+  'simpleChatBase'
+  'copy:libsForSimpleChatWebApp'
+]
+
+taskManager.add 'simpleChat', [
+  'simpleChatChromeApp'
+  'simpleChatWebApp'
 ]
 
 # Build the copy/paste freedom chat sample app.
@@ -94,20 +108,6 @@ taskManager.add 'simpleTurn', [
   'browserify:turnFrontendFreedomModule'
   'copy:libsForSimpleTurnChromeApp'
   'copy:libsForSimpleTurnFirefoxApp'
-]
-
-taskManager.add 'simpleChurnChatChromeApp', [
-  'base'
-  'browserify:simpleChurnChatFreedomModule'
-  'browserify:simpleChurnChatChromeApp'
-  'copy:libsForSimpleChurnChatChromeApp'
-]
-
-taskManager.add 'copypasteChurnChatChromeApp', [
-  'base'
-  'browserify:copypasteChurnChatFreedomModule'
-  'browserify:copypasteChurnChatChromeApp'
-  'copy:libsForCopyPasteChurnChatChromeApp'
 ]
 
 taskManager.add 'adventureBase', [
@@ -399,14 +399,27 @@ module.exports = (grunt) ->
           pathsFromDevBuild: ['simple-turn', 'turn-frontend', 'turn-backend', 'loggingprovider']
           localDestPath: 'samples/simple-turn-firefoxapp/data'
 
-      libsForSimpleChurnChatChromeApp:
+      libsForSimpleChatChromeApp:
         Rule.copyLibs
           npmLibNames: ['freedom-for-chrome']
-          pathsFromDevBuild: ['churn-pipe', 'loggingprovider']
+          pathsFromDevBuild: ['simple-chat', 'churn-pipe', 'loggingprovider']
           pathsFromThirdPartyBuild: [
             'freedom-port-control'
           ]
-          localDestPath: 'samples/simple-churn-chat-chromeapp/'
+          localDestPath: 'samples/simple-chat-chromeapp/'
+      # Simple chat.
+      # While neither churn-pipe nor freedom-port-control can be used in a
+      # regular web page environment, they are included so that obfuscation
+      # may be easily enabled in the Chrome and Firefox samples.
+      libsForSimpleChatWebApp:
+        Rule.copyLibs
+          npmLibNames: ['freedom']
+          pathsFromDevBuild: ['simple-chat', 'churn-pipe', 'loggingprovider']
+          pathsFromThirdPartyBuild: [
+            'freedom-port-control'
+          ]
+          localDestPath: 'samples/simple-chat-webapp/'
+
       libsForCopyPasteChurnChatChromeApp:
         Rule.copyLibs
           npmLibNames: ['freedom-for-chrome']
@@ -553,7 +566,7 @@ module.exports = (grunt) ->
       simpleTurnFreedomModule: Rule.browserify 'simple-turn/freedom-module'
       turnBackendFreedomModule: Rule.browserify 'turn-backend/freedom-module'
       turnFrontendFreedomModule: Rule.browserify 'turn-frontend/freedom-module'
-      simpleChurnChatFreedomModule: Rule.browserify 'samples/simple-churn-chat-chromeapp/freedom-module'
+      simpleChatFreedomModule: Rule.browserify 'simple-chat/freedom-module'
       copypasteChurnChatFreedomModule: Rule.browserify 'samples/copypaste-churn-chat-chromeapp/freedom-module'
       adventureFreedomModule: Rule.browserify 'adventure/freedom-module'
       uprobeFreedomModule: Rule.browserify 'uprobe/freedom-module'
@@ -602,10 +615,8 @@ module.exports = (grunt) ->
       # Browserify sample apps main freedom module and core environments
       copypasteFreedomChatFreedomModule: Rule.browserify 'samples/copypaste-freedom-chat/freedom-module'
       copypasteFreedomChatMain: Rule.browserify 'samples/copypaste-freedom-chat/main.core-env'
-      simpleFreedomChatFreedomModule: Rule.browserify 'samples/simple-freedom-chat/freedom-module'
-      simpleFreedomChatMain: Rule.browserify 'samples/simple-freedom-chat/main.core-env'
+      simpleChatMain: Rule.browserify 'simple-chat/main.core-env'
       copypasteSocksMain: Rule.browserify 'copypaste-socks/main.core-env'
-      simpleChurnChatChromeApp: Rule.browserify 'samples/simple-churn-chat-chromeapp/main.core-env'
       copypasteChurnChatChromeApp: Rule.browserify 'samples/copypaste-churn-chat-chromeapp/main.core-env'
       # Integration tests.
       integrationTcpFreedomModule:
