+++ src/samples/simple-chat-chromeapp/background.js
