+++ src/samples/simple-chat-webapp/main.html
@@ -1,7 +1,8 @@
 <html>
 <head>
-  <title>churn chat client</title>
-  <script src='freedom-for-chrome/freedom-for-chrome.js'></script>
+  <title>Simple chat</title>
+  <script src='freedom/freedom.js'></script>
+  <script src='glue.js'></script>
 </head>
 
 <body>
@@ -27,6 +28,6 @@
     <button id="stopButton">Stop</button>
   </div>
 
-  <script src='main.core-env.static.js'></script>
+  <script src='uproxy-lib/simple-chat/main.core-env.static.js'></script>
 </body>
 </html>
