--- src/samples/simple-churn-chat-chromeapp/message.types.d.ts
@@ -1,5 +1,0 @@
-interface Message {
-  message :string;
-}
-
-export = Message;
