--- src/samples/simple-freedom-chat/freedom-module.json
@@ -1,22 +1,0 @@
-{
-  "name": "simple freedom chat",
-  "app": {
-    "script": [
-      "freedom-module.static.js"
-    ]
-  },
-  "dependencies": {
-    "loggingcontroller": {
-      "url": "uproxy-lib/loggingprovider/freedom-module.json",
-      "api": "loggingcontroller"
-    },
-    "logginglistener": {
-      "url": "uproxy-lib/loggingprovider/freedom-module.json",
-      "api": "logginglistener"
-    }
-  },
-  "permissions": [
-    "core.rtcdatachannel",
-    "core.rtcpeerconnection"
-  ]
-}
