+++ src/simple-chat/freedom-module.json
@@ -1,6 +1,5 @@
 {
-  "name": "simple churn chat",
-  "description": "top-level freedom manifest",
+  "name": "simple chat",
   "app": {
     "script": [
       "freedom-module.static.js"
@@ -8,24 +7,23 @@
   },
   "dependencies": {
     "loggingcontroller": {
-      "url": "uproxy-lib/loggingprovider/freedom-module.json",
+      "url": "../loggingprovider/freedom-module.json",
       "api": "loggingcontroller"
     },
     "logginglistener": {
-      "url": "uproxy-lib/loggingprovider/freedom-module.json",
+      "url": "../loggingprovider/freedom-module.json",
       "api": "logginglistener"
     },
     "churnPipe": {
-      "url": "uproxy-lib/churn-pipe/freedom-module.json",
+      "url": "../churn-pipe/freedom-module.json",
       "api": "churnPipe"
     },
     "portControl": {
-      "url": "freedom-port-control/dist/port-control.json",
+      "url": "../../freedom-port-control/dist/port-control.json",
       "api": "portControl"
     }
   },
   "permissions": [
-    "core.udpsocket",
     "core.rtcdatachannel",
     "core.rtcpeerconnection"
   ]
