+++ src/simple-chat/main.core-env.ts
@@ -1,6 +1,4 @@
-/// <reference path='../../../../third_party/typings/freedom/freedom-core-env.d.ts' />
-
-import Message = require('./message.types');
+/// <reference path='../../../third_party/typings/freedom/freedom-core-env.d.ts' />
 
 var sendButtonA = document.getElementById("sendButtonA");
 var sendButtonB = document.getElementById("sendButtonB");
@@ -12,14 +10,10 @@ var receiveAreaB = <HTMLInputElement>document.getElementById("receiveAreaB");
 
 var stopButton = document.getElementById("stopButton");
 
-freedom('freedom-module.json', {
-    'logger': 'uproxy-lib/loggingprovider/freedom-module.json',
-    'debug': 'debug'
-  }).then(
-    (simpleChatFactory:() => freedom.OnAndEmit<any,any>) => {
-  // TODO: typings for the freedom module
-  var chat :freedom.OnAndEmit<any,any> = simpleChatFactory();
+// Platform-specific function to load the freedomjs module (glue.js).
+declare var loadModule: () => Promise<freedom.OnAndEmit<any, any>>;
 
+loadModule().then((chat:freedom.OnAndEmit<any,any>) => {
   chat.on('ready', function() {
     sendAreaA.disabled = false;
     sendAreaB.disabled = false;
@@ -31,15 +25,13 @@ freedom('freedom-module.json', {
   });
 
   function send(suffix:string, textArea:HTMLInputElement) {
-    chat.emit('send' + suffix, {
-      message: textArea.value || '(empty message)'
-    });
+    chat.emit('send' + suffix, textArea.value || '(empty message)');
   }
   sendButtonA.onclick = send.bind(null, 'A', sendAreaA);
   sendButtonB.onclick = send.bind(null, 'B', sendAreaB);
 
-  function receive(textArea:HTMLInputElement, msg:Message) {
-    textArea.value = msg.message;
+  function receive(textArea:HTMLInputElement, msg:string) {
+    textArea.value = msg;
   }
   chat.on('receiveA', receive.bind(null, receiveAreaA));
   chat.on('receiveB', receive.bind(null, receiveAreaB));
@@ -48,5 +40,5 @@ freedom('freedom-module.json', {
     chat.emit('stop');
   };
 }, (e:Error) => {
-  console.error('could not load freedom: ' + e.message);
+  console.error('could not load freedom module: ' + e.message);
 });
